// frontends/epro/src/app/api/model/dataset-dto.ts
/**
 * KV API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: DEV-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DatasetModelDTO } from './dataset-model-dto';
import { FieldDTO } from './field-dto';


/**
 * Dataset
 */
export interface DatasetDTO { 
    /**
     * Scope reference
     */
    scopePk: number;
    scopeId: string;
    scopeCodeAndShortname: string;
    /**
     * Event reference
     */
    eventPk?: number;
    eventId?: string;
    eventShortname?: string;
    /**
     * Dataset reference
     */
    pk: number;
    /**
     * Dataset ID
     */
    id: string;
    creationTime: Date;
    lastUpdateTime: Date;
    removed: boolean;
    /**
     * Is the dataset referenced by a removed object?
     */
    inRemoved: boolean;
    /**
     * Does the user have the rights to write to the dataset?
     */
    canWrite: boolean;
    /**
     * Can the user remove the dataset?
     */
    canBeRemoved: boolean;
    model: DatasetModelDTO;
    /**
     * ID of the dataset model
     */
    modelId: string;
    /**
     * Fields associated with the dataset
     */
    fields: Array<FieldDTO>;
    /**
     * User-provided rationale for dataset deletion
     */
    rationale?: string;
}