// frontends/epro/src/app/api/model/dataset-update-dto.ts
/**
 * KV API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: DEV-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FieldUpdateDTO } from './field-update-dto';


/**
 * Dataset update
 */
export interface DatasetUpdateDTO { 
    /**
     * Dataset reference
     */
    pk: number;
    /**
     * Fields associated with the dataset
     */
    fields: Array<FieldUpdateDTO>;
}