// frontends/epro/src/app/api/model/field-update-dto.ts
/**
 * KV API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: DEV-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Field update
 */
export interface FieldUpdateDTO { 
    /**
     * Field reference
     */
    pk?: number;
    /**
     * Field model ID
     */
    modelId: string;
    /**
     * Field value
     */
    value: string;
    /**
     * Field file reference
     */
    filePk?: number;
    reset?: boolean;
    rationale?: string;
}